Bug 8868: ILS-DI: CancelHold needs to take a reserve_id
CancelHold takes two parameters: patron_id and item_id.
If item_id is considered as an itemnumber, holds on title can't be
canceled.
If item_id is considered as a biblionumber, all holds on this
biblionumber (for a borrower) will be canceled.

So CancelHold have to consider item_id as a reserve_id.

- Added subroutine C4::Reserves::GetReserve
- C4::ILSDI::Services::GetRecords now returns the reserve_id
- Fix the text in the ilsdi.pl?service=Describe&verb=CancelHold page
- Unit tests for CancelReserved and GetReserve
- Do not delete row in reserves table if insert in old_reserves fails


Signing off, while noting a style issue in the patch review

Passes tests and QA script.
Placed and cancelled a hold using ILS-DI successfully.
Adding a follow-up to also update the ils-di documentation
page in the bootstrap theme.


EDIT: I removed the changes it did to the prog theme.

+++ C4/Reserves.pm
@@ -1063,34 +1063,37 @@ sub CancelReserve {
     my $dbh = C4::Context->dbh;
 
     my $reserve = GetReserve( $reserve_id );
+    if ($reserve) {
+        my $query = "
+            UPDATE reserves
+            SET    cancellationdate = now(),
+                   found            = Null,
+                   priority         = 0
+            WHERE  reserve_id = ?
+        ";
+        my $sth = $dbh->prepare($query);
+        $sth->execute( $reserve_id );
 
-    my $query = "
-        UPDATE reserves
-        SET    cancellationdate = now(),
-               found            = Null,
-               priority         = 0
-        WHERE  reserve_id = ?
-    ";
-    my $sth = $dbh->prepare($query);
-    $sth->execute( $reserve_id );
+        $query = "
+            INSERT INTO old_reserves
+            SELECT * FROM reserves
+            WHERE  reserve_id = ?
+        ";
+        $sth = $dbh->prepare($query);
+        $sth->execute( $reserve_id );
 
-    $query = "
-        INSERT INTO old_reserves
-        SELECT * FROM reserves
-        WHERE  reserve_id = ?
-    ";
-    $sth = $dbh->prepare($query);
-    $sth->execute( $reserve_id );
+        $query = "
+            DELETE FROM reserves
+            WHERE  reserve_id = ?
+        ";
+        $sth = $dbh->prepare($query);
+        $sth->execute( $reserve_id );
 
-    $query = "
-        DELETE FROM reserves
-        WHERE  reserve_id = ?
-    ";
-    $sth = $dbh->prepare($query);
-    $sth->execute( $reserve_id );
+        # now fix the priority on the others....
+        _FixPriority({ biblionumber => $reserve->{biblionumber} });
+    }
 
-    # now fix the priority on the others....
-    _FixPriority({ biblionumber => $reserve->{biblionumber} });
+    return $reserve;
 }
 
 =head2 ModReserve
